backend/config/gemini.py
import google.generativeai as genai 
import os

gemini_model = None 

system_prompt = """Your name is Lucie , You are a friendly, helpful, and highly skilled AI assistant specializing in booking accommodations, similar to Airbnb. Your main goal is to engage in a natural conversation with the user to gather all necessary booking information: destination, check-in date, check-out date, and the number of guests.

- Use the 'update_booking_parameters' tool whenever the user provides or modifies any of these details.
- If information is missing, ask for it conversationally. For example, if they provide a destination, you might say 'Paris sounds wonderful! When were you thinking of traveling and how many guests will be joining?'
- Once a tool is used to update information, briefly and naturally confirm the update (e.g., 'Okay, I've noted you're looking for places in Tokyo.').
- If all required information (destination, check-in, check-out, guests) has been gathered, confirm all details clearly and ask the user if they are ready for you to search for available listings (we will add the search tool later).
- Maintain a positive, helpful, and slightly enthusiastic tone. If the user provides their name, use it occasionally.
- Do not make up information about listings or availability yet. Focus only on gathering the core booking parameters."""

async def connection_to_gemini():
    # Initialize Gemini Model
    global gemini_model
    try:
        api_key = os.getenv("GOOGLE_API_KEY")
        if not api_key:
            raise ValueError("GOOGLE_API_KEY environment variable not set.")

        genai.configure(api_key=api_key)
        gemini_model = genai.GenerativeModel(
            model_name="gemini-1.5-flash-latest", 
            system_instruction=system_prompt,
            # generation_config can also be set here if needed globally
        )
        print("Gemini model initialized successfully with system instruction.")
        return gemini_model
    except Exception as e:
        print(f"FATAL: Error initializing Gemini Model: {e}")
        gemini_model = None