backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "google-genai>=1.10.0",
    "google-generativeai>=0.8.4",
    "redis>=5.2.1",
]
